requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.1.3
django-crispy-forms==1.14.0
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7.1
requests==2.29.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
